refactor(rust): Phase 0 — thin cdylib redesign + kernel as rlib-only
Breaks the kernel ↔ {parallel-crate} cdylib registration cycle that
deferred Phases D-rest / E / F / G in #3916. Mechanism: a dedicated
\`rust/nexus-cdylib/\` crate now owns the sole \`#[pymodule] fn
nexus_kernel\` across the workspace and pulls together the rlibs that
compose the runtime — kernel, lib, raft (and post-Phase-2/3/4 also
backends, services, transport). Linux's \`make bzImage\` analogue: the
cdylib is a build artifact, not an architectural tier.

After this commit the dependency graph has zero cycles:

    contracts → lib → kernel → {backends, services, transport, raft}
                                              ↓
                                        nexus-cdylib (sink)

Concrete changes
----------------
* New crate \`rust/nexus-cdylib/\`:
  - \`[lib] name = "nexus_kernel"  crate-type = ["cdylib"]\`
  - Sole crate that enables \`pyo3/extension-module\` (Cargo unifies
    pyo3 features so transitive consumers see the cdylib-correct
    linkage shape: pyo3 does NOT link libpython itself).
  - \`src/lib.rs\` body is a 4-line \`#[pymodule]\` that calls
    \`lib::python::register\`, \`kernel::python::register\`, and
    \`nexus_raft::pyo3_bindings::register_python_classes\`.
  - \`pyproject.toml\` migrated from \`rust/kernel/pyproject.toml\` —
    same wheel name (\`nexus-kernel\` → \`nexus_kernel-*.whl\`), same
    \`module-name = "nexus_kernel"\`. Wheel filename / Python import
    surface unchanged.

* \`rust/kernel/Cargo.toml\`:
  - \`[lib] crate-type = ["cdylib", "rlib"]\` → \`["rlib"]\`
  - \`[lib] name = "nexus_kernel"\` → \`"kernel"\` (the cdylib name
    belongs to nexus-cdylib now; package = lib name = "kernel" again).
  - Drop \`extension-module\` from kernel's pyo3 features — rlibs
    must NOT enable it themselves (would break \`cargo test\` on the
    rlib alone, and pollutes other consumers).

* \`rust/kernel/src/lib.rs\`:
  - \`#[pymodule] fn nexus_kernel\` body lifted into the new
    \`kernel/src/python.rs\` (\`pub fn register\`) so the cdylib in
    a separate crate can call it.
  - \`use pyo3::prelude::*\` at lib.rs root removed (no longer
    needed — \`python.rs\` brings its own).

* Workspace \`Cargo.toml\`:
  - Add \`rust/nexus-cdylib\` to members list.
  - Update \`kernel\` / \`backends\` / \`raft\` workspace-dep commentary
    to reflect the new role split (kernel is rlib-only; nexus-cdylib
    is the sole cdylib build artifact; \`default-features = false\` on
    kernel keeps runtime defaults from leaking into peer rlibs).

* \`scripts/codegen_kernel_abi.py\`:
  - \`parse_lib_exports\` now scans \`kernel/src/python.rs\` alongside
    \`lib.rs\` and \`lib/src/python/mod.rs\` so the \`add_class::<…>\` /
    \`wrap_pyfunction!\` registrations stay visible to the generator
    after the move.
  - python.rs imports identifiers via \`use crate::…\` (rather than
    fully-qualified \`crate::shm_pipe::Foo\` paths) because the
    existing \`add_class::<MOD::Name>\` regex captures exactly two
    \`::\`-separated segments — a 3-segment path would silently drop
    out of the generated stubs.

Verification
------------
* \`cargo check --workspace --tests\` clean
* \`cargo clippy -p nexus-cdylib -p kernel --tests -- -D warnings\` clean
* \`cargo fmt --check\` clean across all workspace crates
* \`scripts/codegen_kernel_abi.py --check\` reports OK on all 5
  generated artifacts (no drift)
* \`maturin build -m rust/nexus-cdylib/Cargo.toml\` produces
  \`nexus_kernel-0.9.43-cp3XX-…whl\` at the expected path
* \`pip install\` + \`import nexus_kernel\` smoke test exposes
  Kernel / OperationContext / VolumeEngine / VFSSemaphore /
  VfsGrpcServerHandle / Metastore / TofuTrustStore / BloomFilter /
  grep_bulk / glob_match_bulk / canonicalize_path

Out of scope (future commits)
-----------------------------
* Build surfaces (Dockerfile / CI workflows / justfile / packaging
  test) still point at \`rust/kernel/Cargo.toml\` — separate commit
  retargets them at \`rust/nexus-cdylib/Cargo.toml\`.
