+++ rust/nexus-cdylib/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "nexus-cdylib"
+version = "0.10.0"
+edition = "2021"
+description = "Python entry-point cdylib (produces nexus_kernel.so / nexus_kernel.pyd). Composes the kernel + lib + raft (and later backends/services/transport) rlibs into the single shared object Python imports as `nexus_kernel`. Build artifact only — not an architectural tier; see KERNEL-ARCHITECTURE.md §6.1."
+
+[lib]
+# Wheel module name stays `nexus_kernel` so existing
+# `from nexus_kernel import …` imports keep working unchanged.
+name = "nexus_kernel"
+crate-type = ["cdylib"]
+
+[dependencies]
+# kernel sits behind workspace `default-features = false` (so
+# `backends`, `services`, etc. depending on the rlib do not pull
+# kernel runtime defaults like the global mimalloc allocator).
+# nexus-cdylib re-enables those defaults explicitly — this is the
+# only build artifact that owns the `nexus_kernel` runtime.
+kernel = { workspace = true, features = [
+    "mimalloc",
+    "connectors",
+    "py-hook-adapters",
+] }
+lib = { workspace = true, features = ["python"] }
+# `raft` already enters the workspace dep map with
+# `features = ["python", "grpc"]` — nothing to add.
+raft = { workspace = true }
+# Sole crate that enables `extension-module`. Cargo unifies pyo3
+# features across the workspace, so every transitive pyo3 user
+# (kernel, lib, raft, backends, …) sees the cdylib-correct
+# linkage shape: pyo3 does NOT link libpython itself; the .so is
+# loaded by the running Python which provides the symbols.
+pyo3 = { version = "0.27.2", features = ["extension-module"] }
+
+[features]
+# Pass-through for the optional `nostr` backend so Docker / CI
+# `--features nostr` flags keep working at the new entry point.
+nostr = ["kernel/nostr"]
